requirements.txt
aiohttp==3.7.4.post0
alembic==1.7.3
apipkg==1.5
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.4.1
asttokens==2.0.4
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bcrypt==3.2.0
bitarray==1.2.2
black==21.9b0
brownie-token-tester==0.3.2
cached-property==1.5.2
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==3.4.8
cytoolz==0.11.0
dataclassy==0.11.0
dnspython==1.16.0
ecdsa==0.17.0
eip712==0.1.0
email-validator==1.1.3
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.0
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
fastapi==0.68.1
Flask==2.0.1
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
hexbytes==0.2.2
hypothesis==6.23.2
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lru-dict==1.1.7
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
motor==2.5.1
multiaddr==0.0.9
multidict==5.2.0
multitasking==0.0.9
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
numpy==1.21.2
packaging==21.0
pandas==1.3.4
pandas-datareader==0.10.0
parsimonious==0.8.1
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.8.0
psycopg2-binary==2.9.1
py==1.10.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.11.0
pydantic==1.8.2
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.16.0
python-jose==3.3.0
python-multipart==0.0.5
pythx==1.6.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.10.8
requests==2.26.0
rlp==1.2.0
rsa==4.7.2
semantic-version==2.8.5
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
tqdm==4.62.3
typed-ast==1.4.3
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.7
uvicorn==0.15.0
varint==1.0.2
vvm==0.1.0
vyper==0.3.0
wcwidth==0.2.5
web3==5.24.0
websockets==9.1
Werkzeug==2.0.1
wrapt==1.13.1
yarl==1.7.0
yfinance==0.1.63
zipp==3.6.0
